requirements.txt
asgiref==3.4.1
Django==4.0
psycopg2-binary==2.9.2
djangorestframework
sqlparse==0.4.2
